requirements.txt
Django==1.8.3
Werkzeug==0.10.4
beautifulsoup4==4.4.0
bleach==1.4.1
chardet==2.3.0
coverage==3.7.1
django-bleach==0.3.0
django-braces==1.8.0
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-forms-bootstrap==3.0.1
django-model-utils==2.2
django-vanilla-views==1.0.4
feedparser==5.2.0
html5lib==0.999999
lxml==3.4.4
psycopg2==2.6.1
pyflakes==0.8.1
requests==2.7.0
requests-mock==0.6.0
six==1.9.0
speedparser==0.2.0
sqlparse==0.1.15